+++ examples/ethers/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @example/ethers
 
+## 0.0.30
+
+### Patch Changes
+
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.4
+  - pelican-web3-lib-common@1.0.4
+  - pelican-web3-lib-ethers@1.0.4
+  - pelican-web3-lib@1.26.4
+
 ## 0.0.29
 
 ### Patch Changes